requirements.txt.txt
Flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0